hieuedittoi29-03

+++ database/migrations/2021_03_29_075801_insert_column_to_tablecart.php
@@ -15,7 +15,7 @@ class InsertColumnToTablecart extends Migration
     {
         Schema::table('table_cart', function (Blueprint $table) {
             //
-            $table->unsignedBigInteger('prd_id')->nullable();
+            $table->BigInteger('prd_id')->nullable();
             $table->foreign('prd_id')->references('id')->on('table_product');
         });
     }
